+++ internal/app/wwctl/container/copy/main.go
@@ -12,7 +12,7 @@ import (
 func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 
 	if len(args) > 2 {
-		wwlog.Warn("cp only requires 2 arguments but you provided %d arguments. Hence, they will be ignored.", len(args))
+		wwlog.Warn("copy only requires 2 arguments but you provided %d arguments. Hence, they will be ignored.", len(args))
 	}
 
 	cdp := &wwapiv1.ContainerCopyParameter{